requirements.txt
strands-agents==1.0.0
streamlit==1.46.1
boto3==1.39.6
python-dotenv==1.1.1
PyPDF2==3.0.1
Pillow==11.3.0
pandas==2.3.1
numpy==2.2.6
